// frontend-master/frontend-master/src/models/Common/Filters.vue.ts
export interface FiltersVueInterface {
    amount_kg_mask?: string | null;
    amount_kg_from_mask?: string | null;
    amount_kg_from?: number | null;
    amount_kg_to_mask?: string | null;
    amount_kg_to?: number | null;
    docs_type?: string | null;
    docs_number?: string | null;

    date_from?: string | null;
    date_to?: string | null;
}